pyproject.toml
[project]
name = "images"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.95.0",
    "click>=8.3.2",
    "jinja2>=3.1.6",
    "pillow>=12.2.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
mkimg = "scripts.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]
